pyproject.toml
[project]
name = "fast-api-trading-platform"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "alpaca-py>=0.42.1",
    "alpaca-trade-api>=3.2.0",
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "passlib[bcrypt]>=1.7.4",
    "plotly>=6.3.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.1",
    "requests>=2.32.5",
    "streamlit>=1.49.1",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.35.0",
    "websockets>=10.4",
    "yfinance>=0.2.58",
]
